### 2022-04-11 #4313 @editorialbot
[PRE REVIEW]: Open-Source Hypothalamic-ForniX (OSHy-X) Atlases and Segmentation Tool for 3T and 7T
**Submitting author:** @Cadaei-Yuvxvs (<a href="http://orcid.org/0000-0002-6659-365X">Jeryn Chang</a>)
**Repository:** https://github.com/Cadaei-Yuvxvs/OSHy-X
**Branch with paper.md** (empty if default branch):
**Version:** v0.3
