### 2022-05-01 #4313 @emdupre
👋  Hi @a3sha2 @fepegar @jordandekraker @kaczmarj !

Would you be willing to review _Open-Source Hypothalamic-ForniX (OSHy-X) Atlases and Segmentation Tool for 3T and 7T_ for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org/)?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The review takes place on GitHub and focuses on [the software](https://github.com/Cadaei-Yuvxvs/OSHy-X) and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04313/joss.04313/10.21105.joss.04313.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
