### 2022-05-01 #4313 @kaczmarj
Yes — I would be glad to be a reviewer.

Best,
Jakub

> On Apr 30, 2022, at 8:19 PM, Elizabeth DuPre ***@***.***> wrote:
>
> ﻿
> 👋 Hi @a3sha2 @fepegar @jordandekraker @kaczmarj !
>
> Would you be willing to review Open-Source Hypothalamic-ForniX (OSHy-X) Atlases and Segmentation Tool for 3T and 7T for JOSS (the Journal of Open Source Software)?
>
> You can learn more about reviewing for JOSS -- including our conflict of interest policy -- here.
>
> The review takes place on GitHub and focuses on the software and this short paper:
> https://github.com/openjournals/joss-papers/blob/joss.04313/joss.04313/10.21105.joss.04313.pdf
>
> If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.
>
> Thank you !
>
