### 2022-05-01 #4313 @a3sha2
Yes,
I will be glad to review it.

Best,
Az

From: Elizabeth DuPre ***@***.***>
Date: Saturday, April 30, 2022 at 8:19 PM
Cc: Azeez Adebimpe ***@***.***>, Mention ***@***.***>
Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: Open-Source Hypothalamic-ForniX (OSHy-X) Atlases and Segmentation Tool for 3T and 7T (Issue #4313)

👋 Hi @a3sha2<https://github.com/a3sha2> @fepegar<https://github.com/fepegar> @jordandekraker<https://github.com/jordandekraker> @kaczmarj<https://github.com/kaczmarj> !

Would you be willing to review Open-Source Hypothalamic-ForniX (OSHy-X) Atlases and Segmentation Tool for 3T and 7T for JOSS (the Journal of Open Source Software)<https://joss.theoj.org/>?

You can learn more about reviewing for JOSS -- including our conflict of interest policy -- here<https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>.

The review takes place on GitHub and focuses on the software<https://github.com/Cadaei-Yuvxvs/OSHy-X> and this short paper:
https://github.com/openjournals/joss-papers/blob/joss.04313/joss.04313/10.21105.joss.04313.pdf

If you're available to review this work, please let me know and I'll add you as a reviewer. Once I have sufficient reviewers I'll open a dedicated review issue.

Thank you !
