### 2022-05-09 #4368 @emdupre
@a3sha2 thanks for checking on this ! I also noticed the error in displaying the compiled PDF document over the weekend, but there's [no reported error in the GH action logs](https://github.com/openjournals/joss-papers/runs/6338099430?check_suite_focus=true) and the document is now visible. So, I think it was likely transitory ! I'll monitor in case this happens again, though.
