### 2022-05-28 Cadaei-Yuvxvs/OSHy-X#4 @kaczmarj
guidelines for repository contributors
hello, i am reviewing the JOSS submission for this project at https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. this issue is related to my review.

i'd like to suggest adding guidelines for potential contributors to this repository. (please see [github's documentation on this](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)). these guidelines explain how people should contribute to this project.

the github documentation says that these should be included in the contributing.md file.

> - Steps for creating good issues or pull requests.
> - Links to external documentation, mailing lists, or a code of conduct.
> - Community and behavioral expectations.
