### 2022-05-28 Cadaei-Yuvxvs/OSHy-X#5 @kaczmarj
usage outside of containers
hello, i am reviewing the JOSS submission for this project at https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. this issue is related to my review.

is this repository intended to be used only within containers? or should it be possible to use on "bare metal"? if the latter, i noticed that some paths are hard-coded to paths in the container

https://github.com/Cadaei-Yuvxvs/OSHy-X/blob/3c04e72c783086c441bbed751ce0312553dbb4b0/OSHyX/OSHy.py#L396

i am not requesting that this project be installable outside of a container. if the developers wish to only distribute this tool in containers, could the README please be updated to reflect this? also that the code in `OSHy.py` assumes that it is running within the container.
