### 2022-05-28 Cadaei-Yuvxvs/OSHy-X#6 @kaczmarj
description of output files
hello, i am reviewing the JOSS submission for this project at https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. this issue is related to my review.

there is a [description of some output files](https://github.com/Cadaei-Yuvxvs/OSHy-X#output) in the readme, but after running the script, i have some output files that are not described. this is the command-line i ran:

```
singularity run -B $(pwd)/output:/output oshyx_0.3_20211223.sif --target /OSHy/sub-test.nii.gz --outdir /output
```

and these are the output files i see:

```
sub-test_denoised_bias-corrected_cropped_T1w.nii.gz  sub-test_sub-04_cropped_T1w_0_log.txt  sub-test_sub-35_cropped_T1w_7_log.txt
sub-test_fornix.nii.gz                               sub-test_sub-05_cropped_T1w_1_log.txt  sub-test_sub-40_cropped_T1w_8_log.txt
sub-test_hypothalamus.nii.gz                         sub-test_sub-06_cropped_T1w_2_log.txt  sub-test_sub-48_cropped_T1w_9_log.txt
sub-test_Intensity.nii.gz                            sub-test_sub-21_cropped_T1w_3_log.txt  sub-test_TargetMaskImageMajorityVoting_Mask.nii.gz
sub-test_Labels.nii.gz                               sub-test_sub-22_cropped_T1w_4_log.txt  sub-test_TargetMaskImageMajorityVoting.nii.gz
sub-test_mosaic.png                                  sub-test_sub-25_cropped_T1w_5_log.txt  sub-test_volumes.csv
sub-test_resampled_Labels.nii.gz                     sub-test_sub-26_cropped_T1w_6_log.txt
```

could a description of the following files be included?
- `sub-test_Intensity.nii.gz`
- `sub-test_TargetMaskImageMajorityVoting.nii.gz`
- `sub-test_TargetMaskImageMajorityVoting_Mask.nii.gz`
- `*_log.txt`
