### 2022-05-28 Cadaei-Yuvxvs/OSHy-X#7 @kaczmarj
test of tool usage in container
hello, i am reviewing the JOSS submission for this project at https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. this issue is related to my review.

i see the pytest tests to test the functionality of `OSHy.py`, but since this project is used through a container, it would be good to test the behavior in the container. for example, perhaps a sample image could be segmented, and assuming this is deterministic (i don't know if it is), the segmentations can be compared to a known state. this way, any regressions can be caught early.
