### 2022-05-28 Cadaei-Yuvxvs/OSHy-X#8 @kaczmarj
comparison to freesurfer recon-all
hello, i am reviewing the JOSS submission for this project at https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. this issue is related to my review.

i assume that many neurimaging researchers use freesurfer's `recon-all` workflow to segment subcortical structures like the hypothalamus and fornix. given this, i think a comparison with `recon-all` should be included.
