### 2022-05-28 #4368 @kaczmarj
@emdupre - i finished my review in https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168. i made several issues in the OSHy-X repository, and i linked them all here.
