### 2022-05-28 3afdc2c
clarify labels in Labels.nii.gz + fix spelling
this commit clarifies the label IDs in Labels.nii.gz. Previously, the list of labels was rendered with lowercase roman numerals, like `i. Left Hypothalamus`. this commit replaces this with integers 1,2,3,4.

this commit also fixes a spelling error: ressampled -> resampled
