### 2022-06-02 0965768
Added description of files - OSHy-X#6
