### 2022-06-03 378b965
Updated Community with contributing guide
