### 2022-06-03 Cadaei-Yuvxvs/OSHy-X#4 @Cadaei-Yuvxvs
Thank you Jakub for the very helpful suggestion. I have now added a contributing.md file here 011cd364cb01b24a9884e51852713ae749e9c0da.
