### 2022-06-03 Cadaei-Yuvxvs/OSHy-X#5 @kaczmarj
if this project will only be distributed and usable in a container, the "installation" section of the readme might include this wording:

>You have the option of running OSHy-X via NeuroDesk, a Docker container by itself, or a Singularity container. OSHy-X is not designed to run outside of a container.
