### 2022-06-03 Cadaei-Yuvxvs/OSHy-X#5 @Cadaei-Yuvxvs
Hi Jakub, yes - OSHy-X is indeed designed to only be used inside a container. This commit 1ce1c949a83a4e06e818a4ffa887cd54aa2a1525 now includes clarification on this. Cheers!
