### 2022-06-03 Cadaei-Yuvxvs/OSHy-X#8 @Cadaei-Yuvxvs
Thank you for the suggestion, Jakub. The hypothalamus is segmented in Freesurfer's `recon-all` using the method outlined by Billot _et al._. I will update the paper to clarify this. I will also add additional comparisons for fornix segmentation by `recon-all`.
