### 2022-06-03 Cadaei-Yuvxvs/OSHy-X#7 @Cadaei-Yuvxvs
Thank you Jakub. I will add a GitHub Actions workflow to address this.
