+++ .github/workflows/test_oshyx.yml
@@ -1,0 +1,28 @@
+name: Test OSHy-X
+on: [push]
+
+jobs:
+  test-oshyx:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Free up space
+      uses: easimon/maximize-build-space@master
+      with:
+        root-reserve-mb: 40000
+        swap-size-mb: 1024
+        remove-dotnet: 'true'
+        remove-android: 'true'
+        remove-haskell: 'true'
+        overprovision-lvm: 'true'
+    - name: Move docker installation
+      run: |
+          sudo mv /var/lib/docker /home/runner/work/docker
+          sudo ln -s /home/runner/work/docker /var/lib/docker
+          sudo systemctl restart docker
+    - uses: actions/checkout@v2
+    - uses: actions/setup-python@v2
+      with:
+        python-version: 3.8
+    - name: test OSHy-X pipeline
+      run: |
+        /bin/bash tests/run_tests.sh
