### 2022-06-09 690593d
Changed /bin/bash to bash
