+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,4 @@
-on: [push]
+on: [workflow_call]
 
 jobs:
   paper:
