### 2022-06-09 550dba0
Fixed test folder typo and docker syntax
