### 2022-06-09 f3d99cf
moved pip install command to docker generator
