+++ test/run_pytest.sh
@@ -1,7 +1,3 @@
 #!/bin/bash
 
-pip install pytest
-
-python -m pip install --upgrade pip
-pip install -r tmp/requirements.txt
 python -m pytest /OSHy
