### 2022-06-09 53ba360
run pytest tests in /tmp not /OSHy
