+++ test/run_tests.sh
@@ -4,7 +4,7 @@ set -e
 cp -r . /tmp/OSHy-X
 
 container=$( cat /tmp/OSHy-X/README.md | grep "docker pull" | cut -d " " -f 3 | cut -d "\`" -f 1 )
-echo $container
+
 sudo docker pull $container
 
-sudo docker run -v /tmp/OSHy-X/test:/tmp --entrypoint "/bin/bash" $container /tmp/run_pytest.sh
+sudo docker run -v /tmp/OSHy-X/test:/tmp --entrypoint "python" $container -m pytest /tmp/
