### 2022-06-09 ee06012
entrypoint /bin/bash
