+++ test/run_pytest.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python -m pytest /tmp
+python -m pytest /tmp/OSHy-X_test.py
