+++ test/OSHy-X_test.py
@@ -63,12 +63,12 @@ def test_convert_to_bool(boolstrings):
 ])
 def test_OSHy_data_tesla(mocker, tesla):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -85,12 +85,12 @@ def test_OSHy_data_tesla(mocker, tesla):
 ])
 def test_OSHy_data_tesla(mocker, tesla):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -106,12 +106,12 @@ def test_OSHy_data_tesla(mocker, tesla):
 ])
 def test_OSHy_data_weighting(mocker, weighting):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -135,12 +135,12 @@ def test_OSHy_data_weighting(mocker, weighting):
 ])
 def test_OSHy_data_weighting(mocker, weighting, expected_weighting):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -156,12 +156,12 @@ def test_OSHy_data_weighting(mocker, weighting, expected_weighting):
 ])
 def test_OSHy_data_bimodal(mocker, bimodal):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -178,12 +178,12 @@ def test_OSHy_data_bimodal(mocker, bimodal):
 ])
 def test_OSHy_data_bimodal(mocker, bimodal):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -200,12 +200,12 @@ def test_OSHy_data_bimodal(mocker, bimodal):
 ])
 def test_OSHy_data_crop(mocker, crop):
     mocker.patch(
-        'OSHyX.ants.image_read',
+        'OSHy.ants.image_read',
         return_value="An ANTsImage object"
     )
 
     mocker.patch(
-        'OSHyX.glob.glob',
+        'OSHy.glob.glob',
         return_value=["image.nii.gz" for i in range(10)]
     )
 
@@ -223,12 +223,12 @@ def test_OSHy_data_crop(mocker, crop):
 ])
[…]
