### 2022-06-12 9f0af7c
Test end-user pipeline with defaults
