### 2022-06-12 723f3b4
New tests for pipeline features
