+++ test/run_tests.sh
@@ -7,16 +7,16 @@ container=$( cat /tmp/OSHy-X/README.md | grep "docker pull" | cut -d " " -f 3 | 
 
 sudo docker pull $container
 
-out_files=("sub-test_Intensity.nii.gz" "sub-test_denoised_bias-corrected_cropped_T1w.nii.gz" \
-"sub-test_resampled_Labels.nii.gz" "sub-test_sub-21_cropped_T1w_3_log.txt"  \
-"sub-test_sub-35_cropped_T1w_7_log.txt" "sub-test_Labels.nii.gz" "sub-test_fornix.nii.gz" \
-"sub-test_sub-04_cropped_T1w_0_log.txt"  "sub-test_sub-22_cropped_T1w_4_log.txt"  \
-"sub-test_sub-40_cropped_T1w_8_log.txt" "sub-test_TargetMaskImageMajorityVoting.nii.gz" \
-"sub-test_hypothalamus.nii.gz" "sub-test_sub-05_cropped_T1w_1_log.txt" \
-"sub-test_sub-25_cropped_T1w_5_log.txt"  "sub-test_sub-48_cropped_T1w_9_log.txt" \
-"sub-test_TargetMaskImageMajorityVoting_Mask.nii.gz" "sub-test_mosaic.png" \
-"sub-test_sub-06_cropped_T1w_2_log.txt" "sub-test_sub-26_cropped_T1w_6_log.txt" \
-"sub-test_volumes.csv")
+out_files_suffix=("_Intensity.nii.gz" "_denoised_bias-corrected_cropped_T1w.nii.gz" \
+"_resampled_Labels.nii.gz" "_sub-21_cropped_T1w_3_log.txt"  \
+"_sub-35_cropped_T1w_7_log.txt" "_Labels.nii.gz" "_fornix.nii.gz" \
+"_sub-04_cropped_T1w_0_log.txt"  "_sub-22_cropped_T1w_4_log.txt"  \
+"_sub-40_cropped_T1w_8_log.txt" "_TargetMaskImageMajorityVoting.nii.gz" \
+"_hypothalamus.nii.gz" "_sub-05_cropped_T1w_1_log.txt" \
+"_sub-25_cropped_T1w_5_log.txt"  "_sub-48_cropped_T1w_9_log.txt" \
+"_TargetMaskImageMajorityVoting_Mask.nii.gz" "_mosaic.png" \
+"_sub-06_cropped_T1w_2_log.txt" "_sub-26_cropped_T1w_6_log.txt" \
+"_volumes.csv")
 
 echo "[DEBUG] Running Pytest"
 
@@ -26,9 +26,170 @@ echo "[DEBUG] Running default pipeline"
 
 sudo docker run -v /tmp:/tmp $container -t /OSHy/sub-test.nii.gz -o /tmp
 
-for out_file in ${out_files[@]}; do [ -f /tmp/sub-test/${out_file} ] && echo "[DEBUG] Test OK: Regular output file $outfile exists." || echo "[DEBUG] Test FAILED: Regular output file $outfile does not exist."; done
+for out_file in ${out_files_suffix[@]}
+do
+    if [ -f /tmp/sub-test/sub-test_${out_file} ] 
+    then
+        echo "[DEBUG] Test OK: Regular output file (sub-test_$out_file) exists."
+    else
+        echo "[DEBUG] Test FAILED: Regular output file (sub-test_$out_file) does not exist."
+        exit 1
+    fi
+done
 
 seg_vols=$( cat /tmp/sub-test/sub-test_volumes.csv | cut -d "," -f 7 | tail -n 4 )
[…]
