### 2022-06-13 ce92ec7
Added cleanup to free memory
