+++ test/run_tests.sh
@@ -114,55 +114,6 @@ sudo rm -r /tmp/sub-test
 
 docker system prune --force
 
-echo "[DEBUG] Running pipeline with T2w input"
-
-out_T2w_files_suffix=("_Intensity.nii.gz" "_Labels.nii.gz" \
-"_TargetMaskImageMajorityVoting.nii.gz" "_TargetMaskImageMajorityVoting_Mask.nii.gz" \
-"_denoised_bias-corrected_cropped_T2w.nii.gz" "_fornix.nii.gz" \
-"_hypothalamus.nii.gz" "_mosaic.png" \
-"_resampled_Labels.nii.gz" "_sub-04_N4Bias_Denoised_cropped_registered_T2w_0_log.txt" \
-"_sub-05_N4Bias_Denoised_cropped_registered_T2w_1_log.txt" \
-"_sub-06_N4Bias_Denoised_cropped_registered_T2w_2_log.txt" \
-"_sub-21_N4Bias_Denoised_cropped_registered_T2w_3_log.txt" \
-"_sub-22_N4Bias_Denoised_cropped_registered_T2w_4_log.txt" \
-"_sub-25_N4Bias_Denoised_cropped_registered_T2w_5_log.txt" \
-"_sub-26_N4Bias_Denoised_cropped_registered_T2w_6_log.txt" \
-"_sub-35_N4Bias_Denoised_cropped_registered_T2w_7_log.txt" \
-"_sub-40_N4Bias_Denoised_cropped_registered_T2w_8_log.txt" \
-"_sub-48_N4Bias_Denoised_cropped_registered_T2w_9_log.txt" \
-"_volumes.csv")
-
-sudo docker run --memory-swap -1 -v /tmp:/tmp $container -t /OSHy/atlases/3T/sub-05_whole_Warped_FLAIR_T2w_defaced.nii.gz -o /tmp -w T2w
-
-for out_file in ${out_T2w_files_suffix[@]}
-do
-    sub_out_file=sub-test${out_file}
-    if [ -f /tmp/sub-test/${sub_out_file} ] 
-    then
-        echo "[DEBUG] Test OK: Regular output file ($sub_out_file) exists."
-    else
-        echo "[DEBUG] Test FAILED: Regular output file ($sub_out_file) does not exist."
-        exit 1
-    fi
-done
-
-seg_vols=$( cat /tmp/sub-05/sub-05_volumes.csv | cut -d "," -f 7 | tail -n 4 )
-
-for seg_vol in ${seg_vols[@]}
-do 
-    if [ $seg_vol -ge 100 ] && [ $seg_vol -le 700 ] 
-    then 
-        echo "[DEBUG] Test OK: Segmented volume ($seg_vol) is within normal parameters."
-    else 
-        echo "[DEBUG] Test FAILED: Segmented volume ($seg_vol) is not within normal parameters."
-        exit 1
-    fi
-done
-
-sudo rm -r /tmp/sub-05
-
-docker system prune --force
-
 echo "[DEBUG] Running pipeline with 7T input"
 
 out_7T_files_suffix=("_Intensity.nii.gz" "_denoised_bias-corrected_cropped_T1w.nii.gz" \
@@ -179,8 +130,8 @@ sudo docker run --memory-swap -1 $container -t /OSHy/atlases/7T/sub-P0004_whole_
 
 for out_file in ${out_7T_files_suffix[@]}
 do
-    sub_out_file=sub-test${out_file}
-    if [ -f /tmp/sub-test/${sub_out_file} ] 
+    sub_out_file=sub-P0004${out_file}
+    if [ -f /tmp/sub-P0004/${sub_out_file} ] 
     then
[…]
