+++ test/run_tests.sh
@@ -126,7 +126,7 @@ out_7T_files_suffix=("_Intensity.nii.gz" "_denoised_bias-corrected_cropped_T1w.n
 "_mosaic.png" "_sub-H0111_cropped_T1w_2_log.txt" "_sub-P0016_cropped_T1w_9_log.txt"\
 "_sub-P0004_cropped_T1w_6_log.txt" "_volumes.csv")
 
-sudo docker run --memory-swap -1 $container -t /OSHy/atlases/7T/sub-P0004_whole_ses-01_T1w_defaced.nii.gz -o /tmp -x 7
+sudo docker run --memory-swap -1 -v /tmp:/tmp $container -t /OSHy/atlases/7T/sub-P0004_whole_ses-01_T1w_defaced.nii.gz -o /tmp -x 7
 
 for out_file in ${out_7T_files_suffix[@]}
 do
