### 2022-06-13 45bddff
Test now accounts for float comparisons
