+++ README.md
@@ -23,10 +23,10 @@ Follow instructions [here](https://neurodesk.github.io/docs/neurodesktop/getting
 1. Install Docker [here](https://docs.docker.com/get-docker/).
 2. Open a terminal and run:
 
-`docker pull jerync/oshyx_0.4:20220609`
+`docker pull jerync/oshyx_0.4:20220614`
 To pull the container. Or run: 
 
-`docker run --rm -v /path/to/data/folder/:/data/ jerync/oshyx_0.4:20220609 --target /data/input_file.nii.gz --outdir /data/output_directory`
+`docker run --rm -v /path/to/data/folder/:/data/ jerync/oshyx_0.4:20220614 --target /data/input_file.nii.gz --outdir /data/output_directory`
 
 ### Singularity
 1. Install Singularity [here](https://sylabs.io/guides/3.5/user-guide/quick_start.html).
@@ -40,7 +40,7 @@ To pull the container.
 ### Docker and Singularity
 
 ```
-Usage: docker run --rm -v /path/to/data:/data jerync/oshyx_0.4:20220609 
+Usage: docker run --rm -v /path/to/data:/data jerync/oshyx_0.4:20220614 
                [-h] -t TARGET [TARGET ...] -o OUTDIR [-c CROP] [-w WEIGHTING]
                [-d DENOISE] [-f FIELDCORRECTION] [-m MOSAIC] [-x TESLA]
                [-b BIMODAL] [-n NTHREADS]
