+++ test/run_pytest.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-python -m pytest /tmp/OSHy-X_test.py
