+++ .github/workflows/test.yml
@@ -1,19 +1,0 @@
-name: Test code logic
-on: [workflow_call]
-
-jobs:
-  test-oshyx:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v2
-      - name: Setup python
-        uses: actions/setup-python@v1
-        with:
-          python-version: "3.6"
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          pip install -r test/requirements.txt
-      - name: Run pytest
-        run: python -m pytest test/
