+++ JOSS_paper/paper.md
@@ -101,13 +101,13 @@ A summary of the pipeline is illustrated in \autoref{fig:1}. The user can specif
 
 ![Visualisation of segmentation performance between manual segmentation and `JLF`. The hypothalamus is shown in red and the fornix in blue. The first three rows show segmentation in coronal, sagittal and axial planes; a 3D rendering of the structures is illustrated in the fourth row. The difference between `JLF` and manual segmentation illustrates over-segmented (red) and under-segmented (green) areas.\label{fig:2}](../Media/OSHy-X_figure_2.png){ width=70% }
 
-Dice overlaps (\autoref{fig:3}) and (ICC; 2-way fixed-rater mixed effects model with single measurement) between the two raters indicate excellent segmentation accuracy. The left and right hypothalamus received scores of 0.90 (0.66-0.98 CI) and 0.91 (0.68-0.98 CI). The left and right fornix received scores of 0.97 (0.87-0.99 CI) and 0.68 (0.13-0.91 CI).
+Dice overlaps (\autoref{fig:3}) and ICC (Intraclass Correlation; 2-way fixed-rater mixed effects model with single measurement) between the two raters indicate excellent segmentation accuracy. The left and right hypothalamus received scores of 0.90 (0.66-0.98 CI) and 0.91 (0.68-0.98 CI). The left and right fornix received scores of 0.97 (0.87-0.99 CI) and 0.68 (0.13-0.91 CI).
 
 ![Dice overlaps between two raters for the left and right lobes of the hypothalamus and fornix. The median Dice’s coefficient for the left and right hypothalamus is 0.94 (0.01 IQR) and 0.96 (0.03 IQR). The median Dice’s coefficient for the left and right fornix are 0.91 (0.06 IQR) and 0.91 (0.03 IQR).\label{fig:3}](../Media/OSHy-X_figure_3.png){ width=70% }
 
-In comparison to a deep learning method for segmentation [@Billot:2020]  (available in `FreeSurfer` v7.2), we found that `JLF` has higher Dice overlaps with the manual segmentations for both 3T and 7T (\autoref{fig:4}). Additionally, we found that compared to cropped priors, whole-brain priors for `JLF` offers modest benefits to segmentation accuracy at 3T, but significant performance benefits at 7T compared to the deep learning method. While whole brain instead of cropped priors for `JLF` improves segmentation performance, computational time increases prohibitively.
+`FreeSurfer` (v7.2) also offers segmentation of both the hypothalamus and fornix – The `recon-all` command segments the hypothalamus by default [@Billot:2020]. We found that `JLF` has higher Dice overlaps with the manual segmentations at both 3T and 7T (\autoref{fig:4}). Similarly, Dice overlaps for the fornix are significantly higher for JLF at both 3T and 7T (\autoref{fig:4}); however, it should be noted that `FreeSurfer` does not segment the fornix by default.  Additionally, we found that compared to cropped priors, whole-brain priors for `JLF` offers modest benefits to segmentation accuracy at 3T and 7T field strengths. While whole brain instead of cropped priors for `JLF` improves segmentation performance, computational time increases prohibitively.
 
[…]
