+++ JOSS_paper/paper.md
@@ -55,7 +55,7 @@ affiliations:
    index: 8
  - name: CSIRO Health and Biosecurity, The Australian eHealth Research Centre, Brisbane, Australia
    index: 9
-date: 11 April 2022
+date: 19 June 2022
 bibliography: paper.bib
 ---
 
@@ -69,8 +69,8 @@ segmentation for 3T and 7T. The atlases are based on 20 manual segmentations,
 which we demonstrate have high interrater agreement. The versatility of the 
 `OSHy-X` tool allows segmentation and volumetric analysis of the hypothalamus
 and fornix from MRI scans. We also demonstrate that `OSHy-X` segmentation 
-outperforms a deep-learning segmentation method for the hypothalamus 
-[@Billot:2020]. We have previously demonstrated the use of `OSHy-X` on a cohort 
+outperforms FreeSurfer segmentation of the hypothalamus 
+[@Billot:2020] and fornix [@Fischl:2002]. We have previously demonstrated the use of `OSHy-X` on a cohort 
 of 329 non-neurodegenerative control participants and 42 patients with ALS to 
 investigate reduced hypothalamic volume and its association with appetite, 
 hypermetabolism and weight loss [@Chang:2022].
@@ -105,7 +105,7 @@ Dice overlaps (\autoref{fig:3}) and ICC (Intraclass Correlation; 2-way fixed-rat
 
 ![Dice overlaps between two raters for the left and right lobes of the hypothalamus and fornix. The median Dice’s coefficient for the left and right hypothalamus is 0.94 (0.01 IQR) and 0.96 (0.03 IQR). The median Dice’s coefficient for the left and right fornix are 0.91 (0.06 IQR) and 0.91 (0.03 IQR).\label{fig:3}](../Media/OSHy-X_figure_3.png){ width=70% }
 
-`FreeSurfer` (v7.2) also offers segmentation of both the hypothalamus and fornix – The `recon-all` command segments the hypothalamus by default [@Billot:2020]. We found that `JLF` has higher Dice overlaps with the manual segmentations at both 3T and 7T (\autoref{fig:4}). Similarly, Dice overlaps for the fornix are significantly higher for JLF at both 3T and 7T (\autoref{fig:4}); however, it should be noted that `FreeSurfer` does not segment the fornix by default.  Additionally, we found that compared to cropped priors, whole-brain priors for `JLF` offers modest benefits to segmentation accuracy at 3T and 7T field strengths. While whole brain instead of cropped priors for `JLF` improves segmentation performance, computational time increases prohibitively.
+`FreeSurfer` (v7.2) also offers segmentation of both the hypothalamus [@Billot:2020] and fornix [@Fischl:2002] – The `recon-all` command segments the hypothalamus by default. We found that `JLF` has higher Dice overlaps with the manual segmentations at both 3T and 7T (\autoref{fig:4}). Similarly, Dice overlaps for the fornix are significantly higher for JLF at both 3T and 7T (\autoref{fig:4}); however, it should be noted that `FreeSurfer` does not segment the fornix by default.  Additionally, we found that compared to cropped priors, whole-brain priors for `JLF` offers modest benefits to segmentation accuracy at 3T and 7T field strengths. While whole brain instead of cropped priors for `JLF` improves segmentation performance, computational time increases prohibitively.
 
[…]
