+++ JOSS_paper/paper.bib
@@ -17,10 +17,21 @@
    journal = {NeuroImage},
    volume = {223},
    pages = {117287},
-   ISSN = {1053-8119 (Print)
-1053-8119},
+   ISSN = {1053-8119},
    DOI = {10.1016/j.neuroimage.2020.117287},
    year = {2020}
+}
+
+@article{Fischl:2002,
+   author = {B. Fischl and D. H. Salat and E. Busa and M. Albert and M. Dieterich and C. Haselgrove and A. van der Kouwe and R. Killiany and D. Kennedy and S. Klaveness and A. Montillo and N. Makris and B. Rosen and A. M. Dale},
+   title = {Whole Brain Segmentation: Automated Labeling of Neuroanatomical Structures in the Human Brain},
+   journal = {Neuron},
+   volume = {33},
+   number = {3},
+   pages = {341-355},
+   ISSN = {0896-6273},
+   DOI = {10.1016/S0896-6273(02)00569-X},
+   year = {2002}
 }
 
 @misc{Chang:OSF,
