+++ .github/workflows/draft-pdf.yml
@@ -1,4 +1,4 @@
-on: [workflow_call]
+on: [push]
 
 jobs:
   paper:
