### 2022-06-19 Cadaei-Yuvxvs/OSHy-X#8 @Cadaei-Yuvxvs
Dear Jakub, I have now updated our JOSS paper draft to include comparisons with FreeSurfer's segmentation of the hypothalamus and fornix, however both structures are not segmented by default using the `recon-all` command. Instead, separate commands need to be called. Our comparisons indicate that Joint Label Fusion outperforms FreeSurfer for both structures at 3T and 7T field strengths. See https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1159648460
