### 2022-06-19 Cadaei-Yuvxvs/OSHy-X#7 @Cadaei-Yuvxvs
Hi Jakub, a new job has been added to a GitHub actions workflow which tests the various pipelines offered by OSHyX using the container - the same way an end user would use OSHyX: https://github.com/Cadaei-Yuvxvs/OSHy-X/actions/workflows/test_oshyx.yml
