### 2022-06-28 #4368 @Cadaei-Yuvxvs
Dear @a3sha2, do you have any specific feedback about this paper? We have addressed and completed issues opened by @kaczmarj regarding automated tests, functionality documentation, state of the field, community guidelines, and installation. We would be very keen to hear your thoughts and address them to improve our work!
