### 2022-07-02 Cadaei-Yuvxvs/OSHy-X#16 @a3sha2
T2w scan 
@Cadaei-Yuvxvs  is a T2w scan compulsory?
what if there is no T2w?
