### 2022-07-02 #4368 @a3sha2
@Cadaei-Yuvxvs sorry for replying late. I had nothing much to add. Most of the issues I would have raised are already asked by @kaczmarj . I  asked a question on [issues page](https://github.com/Cadaei-Yuvxvs/OSHy-X/issues/16).  I just want to know if only T1w scan will be enough if there is no T2w scan

I can run the docker image completely
welldone to the team

@emdupre
