### 2022-07-02 #4368 @kaczmarj
@emdupre - in my initial checklist review (https://github.com/openjournals/joss-reviews/issues/4368#issuecomment-1116524168), i left some boxes unchecked and made issues for each of those items. the authors have addressed these comments. should i check those boxes now? or is it sufficient to say that i am now satisfied as a reviewer?
