### 2022-07-03 Cadaei-Yuvxvs/OSHy-X#16 @Cadaei-Yuvxvs
Hi @a3sha2 thanks very much for your question. T2w is not compulsory. By default, OSHy-X accepts T1w images. However, the user can use the `--weighting` flag to specify if the input image is T2w instead.
