### 2022-07-03 #4368 @Cadaei-Yuvxvs
Thank you @a3sha2 :) Very kglad that we've raised most of the issues you raised. I've just posted a reply to the issue you opened for OSHy-X. https://github.com/Cadaei-Yuvxvs/OSHy-X/issues/16
