### 2022-07-08 #4368 @emdupre
Thank you, @kaczmarj and @a3sha2 ! Confirming that you've both signed off that the submission meets JOSS's publication requirements in its current form.

@Cadaei-Yuvxvs, I'm going to run a few additional checks to make sure that all of the editorial requirements are also met !
