### 2022-07-15 #4368 @emdupre
Thanks again for your patience here, @Cadaei-Yuvxvs ! A few small asks on the paper itself :

- [ ] Could you please update the text under the `Tool` subsection of `Methodology`  to fix the formatting of the Wang reference; i.e., `Joint Label Fusion (JLF) [@Wang:2013]` --> Joint Label Fusion (JLF; `@Wang:2013`)
- [ ] Please add a reference for `MriResearchTools`, as currently only the 0.5.2 version number is listed.
- [ ] On L65 in `Performance`, please add a comma after "however."
- [ ] In the `Atlas` subsection of `Methodology`, you note that "acquisition parameters have been outlined previously" but point to Chang et al., 2022, which is currently listed as an unpublished work. Has this reference since been published, or is there a place where readers could access the acquisition parameters ?
- [ ] Figure 4 caption: It's currently unclear what the "deep learning method" refers to, as your citations currently only point to the Billot (2020) reference as a deep learning method (`Statement of Need`) but suggest that it only applies to the hypothalamus (`Performance`). Could you please clarify the figure caption ?
- [ ] Update the journal on the Wolff reference from "Psychiatry research. Neuroimaging" to "Psychiatry Research: Neuroimaging".

After making these edits, could you then please :

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare or an institutional repository)
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with accepting the submission 🚀
