### 2022-07-21 #4368 @Cadaei-Yuvxvs
Thanks very much @emdupre! I have made the edits requested.

In regards to the acquisition parameters - I have updated the reference to our atlas repository instead, which now has the parameters in its wiki. Our unpublished work is currently undergoing revisions with the European Journal of Neurology.

Let me know if there's anything else that needs editing 😊

I have made a tagged release: OSHy-X v0.4: https://github.com/Cadaei-Yuvxvs/OSHy-X/releases/tag/0.4
The archive is now uploaded to Zenodo and I have checked that the metadata is correct.
[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.6874987.svg)](https://doi.org/10.5281/zenodo.6874987)
