### 2022-07-25 #4368 @emdupre
@openjournals/joss-eics I think this article is ready for `recommend-accept` but the PDF compilation is silently failing. Any suggestions how to approach debugging this, or if the issue is more likely on the editorial infra ?
