### 2022-08-02 #4368 @Cadaei-Yuvxvs
@emdupre Thank you for following this up! The PDF compilation appears to be working now :)
