### 2022-08-03 #4368 @emdupre
Sorry, in reviewing the final draft it looks like one small point still needs to be addressed :

> Could you please update the text under the Tool subsection of Methodology to fix the formatting of the Wang reference; i.e., Joint Label Fusion (JLF) [@Wang:2013] --> Joint Label Fusion (JLF; @Wang:2013)

The same issue would also apply to the (new !) `MriResearchTools (v0.5.2; [@Eckstein:2022])` reference. In both cases, you should be able to drop the square brackets entirely.
