### 2022-08-03 4677743
Drop square brackets for refs in parathenses
