### 2022-08-03 #4368 @Cadaei-Yuvxvs
@emdupre Apologies for that. I've now dropped the square brackets for these two references in `paper.md`.  The references now look like

> Joint Label Fusion (JLF; Wang et al. (2013))
