### 2022-08-03 #4368 @emdupre
Thank you, @Cadaei-Yuvxvs !

I'm now happy to recommend OSHy-X to the EIC team for publication -- congratulations on this impressive effort !
