### 2022-08-05 #4368 @Cadaei-Yuvxvs
> Thank you, @Cadaei-Yuvxvs !
>
> I'm now happy to recommend OSHy-X to the EIC team for publication -- congratulations on this impressive effort !

Thank you @emdupre 😊 That's great news!
