### 2022-08-05 #4368 @arfon
@a3sha2 @emdupre – it looks like @a3sha2 didn't check off the 'substantial scholarly effort' box here. Could you confirm that you're happy to check this off?
